feat: rag demo
